smartsupport_ai/requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
chromadb
google-generativeai
python-dotenv
pydantic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
prometheus_client
redis
